requirements.txt
python-telegram-bot==21.10
google-generativeai==0.8.3
flask==3.1.0
